Textures updated
Textures have been updated for: Mercury, Venus, Earth and Moon.
Also, we have created a file for each graphic quality, so the
texture size to quality correspondence is:
low - 1K
normal - 2K
high - 4K or 8K

In the future we will add an "Ultra" mode which will
get the 8K textures, and the High will get 4K.
This fixes #82.