Disabled large eye separations in 3D mode
3D mode uses now the approach of updating the scene
once and rendering twice.
This presents problems with large eye separations, due
to the camera being floated, namely:
-  They run into float precision jittering
-  Misdetections in LOD

To fix this we need to either get rid completely of
primitives (points, lines, etc.) and use only triangle
based rendering or do a two step approach where we update
and render the whole scene twice for a single 3D frame. This
second approach would kill the FPS.