Add controls.invert_focus.xy option
This adds the option to invert the response of joystick pitch and yaw
motion in focus mode. Default value "false" keeps previous behaviour,
"true" inverts it. Rationale: joystick motion in focus mode so far IMHO
had inversed logic compared to free and space ship mode. If you agree
with this opinion, the option makes this possible, having the camera
move to the left when pushing the joystick to the left and having the
camera move upward when when pulling the joystick.